requirements.txt
aioapns==3.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
amqp==5.3.1
asgiref==3.9.1
async-timeout==5.0.1
attrs==25.3.0
billiard==4.2.1
cachetools==5.5.2
celery==5.5.3
certifi==2025.8.3
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
crispy-bootstrap4==2025.6
cron_descriptor==2.0.5
cryptography==45.0.7
distlib==0.4.0
Django==5.2.5
django-admin-interface==0.30.1
django-allauth==65.11.1
django-celery-beat==2.8.1
django-colorfield==0.14.0
django-crispy-forms==2.4
django-environ==0.12.0
django-hijack==3.7.4
django-jazzmin==3.0.1
django-timezone-field==7.1
django-walletpass==4.2.1
django-widget-tweaks==1.5.0
djangorestframework==3.16.1
fido2==2.0.0
filelock==3.19.1
frozenlist==1.7.0
google-api-core==2.25.1
google-api-python-client==2.179.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
h2==4.3.0
hpack==4.1.0
httplib2==0.30.0
hyperframe==6.1.0
identify==2.6.13
idna==3.10
iniconfig==2.1.0
kombu==5.5.4
Markdown==3.8.2
multidict==6.6.4
nodeenv==1.9.1
oauthlib==3.3.1
packaging==25.0
pillow==11.3.0
platformdirs==4.4.0
pluggy==1.6.0
pre_commit==4.3.0
prompt_toolkit==3.0.52
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.32.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
Pygments==2.19.2
PyJWT==2.10.1
pyOpenSSL==25.1.0
pyparsing==3.2.3
pytest==8.4.1
pytest-django==4.11.1
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-slugify==8.0.4
PyYAML==6.0.2
qrcode==8.2
redis==6.4.0
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
spond==1.1.1
sqlparse==0.5.3
text-unidecode==1.3
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
virtualenv==20.34.0
wcwidth==0.2.13
yarl==1.20.1
